+++ libs/polysemy-wire-zoo/polysemy-wire-zoo.cabal
@@ -21,6 +21,8 @@ library
     Wire.Sem.Now.Input
     Wire.Sem.Now.IO
     Wire.Sem.Now.Spec
+    Wire.Sem.Paging
+    Wire.Sem.Paging.Cassandra
     Wire.Sem.Random
     Wire.Sem.Random.IO
 
@@ -74,6 +76,7 @@ library
       base             >=4.6 && <5.0
     , HsOpenSSL
     , hspec
+    , cassandra-util
     , imports
     , polysemy
     , polysemy-check
@@ -84,5 +87,6 @@ library
     , tinylog
     , types-common
     , uuid
+    , wire-api
 
   default-language:   Haskell2010
