+++ libs/polysemy-wire-zoo/src/Wire/Sem/Paging.hs
@@ -15,7 +15,7 @@
 -- You should have received a copy of the GNU Affero General Public License along
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
-module Galley.Effects.Paging
+module Wire.Sem.Paging
   ( -- * General paging types
     Page,
     PagingState,
