+++ services/galley/src/Galley/API/Internal.hs
@@ -49,7 +49,7 @@ import Galley.API.Teams.Features
 import qualified Galley.API.Update as Update
 import Galley.API.Util
 import Galley.App
-import Galley.Cassandra.Paging
+import Wire.Sem.Paging.Cassandra
 import Galley.Cassandra.TeamFeatures
 import qualified Galley.Data.Conversation as Data
 import Galley.Effects
@@ -59,7 +59,7 @@ import Galley.Effects.FederatorAccess
 import Galley.Effects.GundeckAccess
 import Galley.Effects.LegalHoldStore as LegalHoldStore
 import Galley.Effects.MemberStore
-import Galley.Effects.Paging
+import Wire.Sem.Paging
 import Galley.Effects.TeamStore
 import qualified Galley.Intra.Push as Intra
 import Galley.Monad
