+++ services/galley/src/Galley/API/Teams.hs
@@ -87,7 +87,7 @@ import qualified Galley.API.Teams.Notifications as APITeamQueue
 import qualified Galley.API.Update as API
 import Galley.API.Util
 import Galley.App
-import Galley.Cassandra.Paging
+import Wire.Sem.Paging.Cassandra
 import qualified Galley.Data.Conversation as Data
 import Galley.Data.Services (BotMember)
 import Galley.Effects
@@ -98,7 +98,7 @@ import qualified Galley.Effects.GundeckAccess as E
 import qualified Galley.Effects.LegalHoldStore as Data
 import qualified Galley.Effects.ListItems as E
 import qualified Galley.Effects.MemberStore as E
-import qualified Galley.Effects.Paging as E
+import qualified Wire.Sem.Paging as E
 import qualified Galley.Effects.Queue as E
 import qualified Galley.Effects.SearchVisibilityStore as SearchVisibilityData
 import qualified Galley.Effects.SparAccess as Spar
