+++ services/galley/src/Galley/Cassandra/ConversationList.hs
@@ -27,9 +27,8 @@ import Data.Id
 import Data.Qualified
 import Data.Range
 import Galley.Cassandra.Instances ()
-import Galley.Cassandra.Paging
+import Wire.Sem.Paging.Cassandra
 import qualified Galley.Cassandra.Queries as Cql
-import Galley.Cassandra.ResultSet
 import Galley.Cassandra.Store
 import Galley.Effects.ListItems
 import Imports hiding (max)
