--- services/galley/src/Galley/Cassandra/ResultSet.hs
@@ -1,51 +1,0 @@
--- This file is part of the Wire Server implementation.
---
--- Copyright (C) 2022 Wire Swiss GmbH <opensource@wire.com>
---
--- This program is free software: you can redistribute it and/or modify it under
--- the terms of the GNU Affero General Public License as published by the Free
--- Software Foundation, either version 3 of the License, or (at your option) any
--- later version.
---
--- This program is distributed in the hope that it will be useful, but WITHOUT
--- ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
--- FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
--- details.
---
--- You should have received a copy of the GNU Affero General Public License along
--- with this program. If not, see <https://www.gnu.org/licenses/>.
-
-module Galley.Cassandra.ResultSet where
-
-import Cassandra
-import Imports
-
--- We use this newtype to highlight the fact that the 'Page' wrapped in here
--- can not reliably used for paging.
---
--- The reason for this is that Cassandra returns 'hasMore' as true if the
--- page size requested is equal to result size. To work around this we
--- actually request for one additional element and drop the last value if
--- necessary. This means however that 'nextPage' does not work properly as
--- we would miss a value on every page size.
--- Thus, and since we don't want to expose the ResultSet constructor
--- because it gives access to `nextPage`, we give accessors to the results
--- and a more typed `hasMore` (ResultSetComplete | ResultSetTruncated)
-data ResultSet a = ResultSet
-  { resultSetResult :: [a],
-    resultSetType :: ResultSetType
-  }
-  deriving stock (Show, Functor, Foldable, Traversable)
-
--- | A more descriptive type than using a simple bool to represent `hasMore`
-data ResultSetType
-  = ResultSetComplete
-  | ResultSetTruncated
-  deriving stock (Eq, Show)
-
-mkResultSet :: Page a -> ResultSet a
-mkResultSet page = ResultSet (result page) typ
-  where
-    typ
-      | hasMore page = ResultSetTruncated
-      | otherwise = ResultSetComplete
